Relocate capcom command
