+++ cmd/capcom/cmd/root.go
