+++ cmd/capcom/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/poka-yoke/spaceflight/cmd/capcom/cmd"
+
+func main() {
+	cmd.Execute()
+}
