--- mcc/capcom/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/poka-yoke/spaceflight/mcc/capcom/cmd"
-
-func main() {
-	cmd.Execute()
-}
