+++ pkg/capcom/capcom.go
