+++ pkg/capcom/capcom_test.go
