+++ pkg/capcom/doc.go
