+++ pkg/capcom/ec2mock_test.go
