+++ pkg/capcom/graphing.go
