+++ pkg/capcom/search.go
