+++ pkg/capcom/search_test.go
